fix: apply `import.meta.client` transform to all components
